[feature] Allow mentioning matchers mentioned by alternation
For instance, this did not work previously:

$(C:1|pa=lb) | $(C:1|ma=fr) -> $(C:1|pa=ve);

but now does. In order to have a matcher from an alternation count,
it must be present in all branches. That is, the following is still
forbidden:

$(C:1|pa=lb) | a -> $(C:1|pa=ve);

since if "a" is matched, then we don't know what to replace it with.

The tests 09-alternation-match and 10-alternation-match-bogus
test this feature.